Bug 1669597 - Use word-alignment to avoid misaligning the stack pointer. r=iain
Also fixes the callVM trampoline on x64. This now matches the ARM64 and MIPS64
implementations.

Differential Revision: https://phabricator.services.mozilla.com/D93075

UltraBlame original commit: 6c44f42ee22d5f2299cf1da0a24ebb3ca10656f3

+++ js/src/jit/MacroAssembler.cpp
@@ -32,7 +32,7 @@
 #include "jit/Simulator.h"
 #include "js/Conversions.h"
 #include "js/friend/DOMProxy.h"  
-#include "js/ScalarType.h"  
+#include "js/ScalarType.h"       
 #include "vm/ArgumentsObject.h"
 #include "vm/ArrayBufferViewObject.h"
 #include "vm/FunctionFlags.h"  
@@ -1948,7 +1948,8 @@ void MacroAssembler::guardStringToInt32(Register str, Register output,
     bind(&vmCall);
 
     
-    reserveStack(sizeof(int32_t));
+    
+    reserveStack(sizeof(uintptr_t));
     moveStackPtrTo(output);
 
     volatileRegs.takeUnchecked(scratch);
@@ -1976,12 +1977,12 @@ void MacroAssembler::guardStringToInt32(Register str, Register output,
       
       
       
-      addToStackPtr(Imm32(sizeof(int32_t)));
+      addToStackPtr(Imm32(sizeof(uintptr_t)));
       jump(fail);
     }
     bind(&ok);
     load32(Address(output, 0), output);
-    freeStack(sizeof(int32_t));
+    freeStack(sizeof(uintptr_t));
   }
   bind(&done);
 }
